refactor: always export caching functionality -- avoid passing up feature flag conditionals that make the consuming code more complex and harder to understand
