+++ crates/host/src/module/wasmer_wamr.rs
@@ -16,11 +16,18 @@ pub fn get_ios_module_from_file(_path: &Path) -> Result<Module, DeserializeError
     unimplemented!("The feature flag 'wasmer_sys' must be enabled to support compiling wasm");
 }
 
+pub(crate) fn make_engine() -> Engine {
+    Engine::default()
+}
+pub(crate) fn make_runtime_engine() -> Engine {
+    Engine::default()
+}
+
 /// Build an interpreter module from wasm bytes.
 /// The interpreter used (wasm micro runtime) does not support metering
 /// See tracking issue: https://github.com/bytecodealliance/wasm-micro-runtime/issues/2163
 pub fn build_module(wasm: &[u8]) -> Result<Arc<Module>, wasmer::RuntimeError> {
-    let compiler_engine = Engine::default();
+    let compiler_engine = make_engine();
     let res = Module::from_binary(&compiler_engine, wasm);
     let module = res.map_err(|e| wasm_error!(WasmErrorInner::ModuleBuild(e.to_string())))?;
     Ok(Arc::new(module))
